requirements.txt
alembic==1.4.3
astroid==2.4.2
blinker==1.4
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Alembic==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
